requirements.txt
streamlit
openai
python-dotenv
pydantic
faiss-cpu
sentence-transformers
pillow
pytesseract
easyocr
whisper
numpy
pandas
scikit-learn
sympy
soundfile